refactor: divide Specs
- 명세서 admin, franchise 분리
