+++ src/main/java/com/akatsuki/pioms/specs/dto/SpecsDTO.java
@@ -1,21 +1,16 @@
 package com.akatsuki.pioms.specs.dto;
 
-import com.akatsuki.pioms.franchise.aggregate.Franchise;
-import com.akatsuki.pioms.franchise.dto.FranchiseDTO;
-import com.akatsuki.pioms.order.aggregate.Order;
 import com.akatsuki.pioms.order.dto.OrderDTO;
 import com.akatsuki.pioms.specs.aggregate.SpecsEntity;
-import jakarta.persistence.Column;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.OneToOne;
+import lombok.AllArgsConstructor;
 import lombok.Getter;
-import lombok.Setter;
+import lombok.NoArgsConstructor;
 
 import java.time.LocalDateTime;
 
 @Getter
-@Setter
+@NoArgsConstructor
+@AllArgsConstructor
 public class SpecsDTO {
     private int specsCode;
     private LocalDateTime specsDate;
