+++ src/main/java/com/akatsuki/pioms/specs/repository/SpecsRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface SpecsRepository extends JpaRepository<SpecsEntity,Integer> {
-    List<SpecsEntity> findAllByFranchiseFranchiseCode(int franchiseCode);
+    List<SpecsEntity> findAllByOrderFranchiseFranchiseCode(int franchiseCode);
 
 //    List<SpecsEntity> findAllAndOrderBySpecsDateDesc();
